requirements.txt
asgiref==3.4.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-adminlte3==0.1.6
django-braces==1.15.0
django-model-utils==4.2.0
django-recaptcha==2.0.6
django-render-partial==0.4
django-star-ratings==0.9.1
django-widget-tweaks==1.4.12
idna==3.3
isodate==0.6.0
lxml==4.6.4
Pillow==8.4.0
platformdirs==2.4.0
python-decouple==3.6
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
swapper==1.2.0
typing-extensions==3.10.0.2
urllib3==1.26.7
zeep==4.1.0
